requirements.txt
dlib==19.24.0
mergedeep==1.3.4
natsort==8.2.0
numpy==1.24.1
opencv-python==4.5.5.64
Pillow==9.3.0
pip-licenses==4.0.3
pyinstaller==5.7.0
tqdm==4.64.1
